Code clean.
