+++ README.md
@@ -1,1 +1,3 @@
-Misie się ścigają!.
+# game-the-care-bear-racing
+
+Misie się ścigają!
