+++ app/assets/maps/test.json
@@ -33,11 +33,11 @@
  "tilesets":[
         {
          "firstgid":1,
-         "image":"tecza.png",
+         "image":"rainbow.png",
          "imageheight":160,
          "imagewidth":40,
          "margin":0,
-         "name":"tecza",
+         "name":"rainbow",
          "properties":
             {
