+++ app/index.html
@@ -1,13 +1,13 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
-	<meta charset="UTF-8">
-	<title>Phaser with a sprinkle of ES6 Dust!</title>
-	<link rel="stylesheet" href="styles/main.css">
+    <meta charset="UTF-8">
+    <title>game-the-care-bear-racing</title>
+    <link rel="stylesheet" href="styles/main.css">
 </head>
 <body>
-	<div id="content"></div>
-	<script src="scripts/vendor/phaser.js"></script>
-	<script src="dist/game.js"></script>
+    <div id="content"></div>
+    <script src="scripts/vendor/phaser.js"></script>
+    <script src="dist/game.js"></script>
 </body>
 </html>
