+++ app/scripts/core/main.es6.js
@@ -1,17 +1,17 @@
-import Menu from './states/MenuState';
+import MenuState from './states/MenuState';
 import GameState from './states/GameState';
 import GameOverState from './states/GameOverState';
 
 class Game extends Phaser.Game {
 
-	constructor() {
-		super(800, 800, Phaser.AUTO, 'content', null);
-		this.state.add('MenuState', Menu);
+    constructor() {
+        super(800, 400, Phaser.AUTO, 'content', null);
+        this.state.add('MenuState', MenuState);
         this.state.add('GameState', GameState);
         this.state.add('GameOverState', GameOverState);
 
-		this.state.start('GameOverState');
-	}
+        this.state.start('MenuState');
+    }
 
 }
