+++ app/styles/main.css
@@ -1,15 +1,15 @@
 html {
-  width: 100%;
-  height: 100%;
-  margin: 0px;
-  padding: 0px;
+    width: 100%;
+    height: 100%;
+    margin: 0px;
+    padding: 0px;
 }
 
 body {
-  height: inherit;
-  width: inherit;
-  margin: 0px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
+    height: inherit;
+    width: inherit;
+    margin: 0px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
 }
