+++ package.json
@@ -1,7 +1,6 @@
 {
   "private": true,
   "dependencies": {
-    "babel-core": "^5.8.22",
     "babel-loader": "^5.3.2",
     "webpack": "^1.11.0"
   }
